requirements.txt
openai
chat-downloader
python-dotenv
requests
sounddevice
soundfile
numpy
openai-whisper
torch
torchaudio